requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
anyio==4.9.0
    # via watchfiles
blinker==1.9.0
    # via pelican
cssmin==0.2.0
    # via -r requirements.in
docutils==0.21.2
    # via pelican
exceptiongroup==1.3.0
    # via anyio
feedgenerator==2.1.0
    # via pelican
idna==3.10
    # via anyio
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   pelican
jsmin==3.0.1
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
ordered-set==4.1.0
    # via pelican
pelican==4.11.0
    # via
    #   -r requirements.in
    #   pelican-webassets
pelican-webassets==2.1.0
    # via -r requirements.in
pygments==2.18.0
    # via
    #   pelican
    #   rich
python-dateutil==2.9.0.post0
    # via pelican
pytz==2025.2
    # via feedgenerator
pyyaml==6.0.2
    # via -r requirements.in
rich==14.0.0
    # via pelican
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
typing-extensions==4.14.0
    # via
    #   anyio
    #   exceptiongroup
    #   rich
unidecode==1.4.0
    # via pelican
watchfiles==1.1.0
    # via pelican
